triangle does not show up

+++ triangles/cuda/include/aabb.h
@@ -55,12 +55,12 @@ public:
 };
 
 __device__ aabb surrounding_box(aabb box0, aabb box1) {
-  point3 small(fmin(box0.min().x(), box1.min().x()),
-               fmin(box0.min().y(), box1.min().y()),
-               fmin(box0.min().z(), box1.min().z()));
-  point3 big(fmax(box0.max().x(), box1.max().x()),
-             fmax(box0.max().y(), box1.max().y()),
-             fmax(box0.max().z(), box1.max().z()));
+  point3 small(fminf(box0.min().x(), box1.min().x()),
+               fminf(box0.min().y(), box1.min().y()),
+               fminf(box0.min().z(), box1.min().z()));
+  point3 big(fmaxf(box0.max().x(), box1.max().x()),
+             fmaxf(box0.max().y(), box1.max().y()),
+             fmaxf(box0.max().z(), box1.max().z()));
   return aabb(small, big);
 }
